# docs/swagger.yaml
basePath: /v1
definitions:
...
  api.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  api.SystemResponse:
    properties:
      data:
        type: object
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: fxDmx REST service for DMX fixture control from Raspberry Pi
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: fxDmx
  version: "1.0"
paths:
...
  /system/usbinfo:
    get:
      consumes:
      - application/json
      description: Gets information about currently connected USB serial devices
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SystemResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Gets information about currently connected USB serial devices
      tags:
      - system
...
swagger: "2.0"